added task "179. Largest Number"

+++ Algorithms/0829.ConsecutiveNumbersSum/solution.cpp
@@ -27,7 +27,7 @@ public:
                 break;
             ++termCount;
         }
-        return caseCount;
+        return static_cast<int>(caseCount);
     }
 };
